Updated build configuration.
Configuration paramaters such as framework search path, header search
path, build product location, and several other folders are now passed
to the Build Framework and Build Extensions scripts instead of being
defined in their configuration files. This is done so that only the
main Textual configuration needs to be updated for the passing of this
information.

Additionally, the actual configuration of each target has been updated
to share the same values. This includes the main Textual executable,
the frameworks, and the extensions.

Build results for everything is now stored in a .tmp folder in the root
directory. The Textual executable result is still located in the
"BuildResults" folder.

+++ Frameworks/Blowfish Encryption/Source/Classes/DH1080Base.mm
@@ -132,7 +132,7 @@ static NSString *fishPrimeB64 = @"++ECLiPSE+is+proud+to+present+latest+FiSH+rele
 	
 	unsigned char *primeBuffer = (unsigned char *)[primeData cStringUsingEncoding:[NSString defaultCStringEncoding]];
 
-	BIGNUM *ret = BN_bin2bn(primeBuffer, primeData.length, self.DHStatus->p);
+	BIGNUM *ret = BN_bin2bn(primeBuffer, (int)primeData.length, self.DHStatus->p);
 
 	DHAssertYES(ret == 0);
 	DHAssertYES(self.DHStatus->g == 0);
@@ -187,7 +187,7 @@ static NSString *fishPrimeB64 = @"++ECLiPSE+is+proud+to+present+latest+FiSH+rele
 
 	unsigned char *publicKeyBuffer = (unsigned char *)[publicKey cStringUsingEncoding:[NSString defaultCStringEncoding]];
 	
-	BIGNUM *ret = BN_bin2bn(publicKeyBuffer, publicKey.length, self.publicBigNum);
+	BIGNUM *ret = BN_bin2bn(publicKeyBuffer, (int)publicKey.length, self.publicBigNum);
 
 	DHAssertYES(ret == 0);
 	DHAssertYES(self.publicBigNum == 0);
